refactoring

+++ Resources/DSC Configurations/OctopusDeploy.ps1
@@ -4,12 +4,15 @@ Configuration OctopusDeploy
         $OctopusNodeName,
         $ConnectionString,
         $HostHeader,
+        $FullyQualifiedUrl,
         $OctopusVersionToInstall = 'latest'
     )
     Import-DscResource -ModuleName PSDesiredStateConfiguration
     Import-DscResource -ModuleName xPSDesiredStateConfiguration
     Import-DscResource -ModuleName xNetworking
     Import-DscResource -ModuleName xSystemSecurity
+
+    $octopusDeployServiceAccount = Get-AutomationPSCredential -Name 'OctopusDeployServiceAccount'
 
     Node Server
     {
@@ -24,7 +27,7 @@ Configuration OctopusDeploy
             Action                = "Allow"
             Profile               = "Any"
             Direction             = "InBound"
-            LocalPort             = ("80", "444", "10943")
+            LocalPort             = ("80", "10943")
             Protocol              = "TCP"
         }
 
@@ -45,38 +48,44 @@ Configuration OctopusDeploy
         {
             Uri = $octopusDownloadUri
             DestinationPath = $octopusInstallFile
+            MatchSource = $false
             DependsOn = '[File]OctopusDeployFolder'
         }
+
         $octopusInstallLogFile = Join-Path $octopusDeployRoot "OctopusServer.$OctopusVersionToInstall.install.log"
-        $octopusInstallStateFile = Join-Path $octopusDeployRoot 'OctopusDeploy.version'
+        $octopusInstallStateFile = Join-Path $octopusDeployRoot 'version.statefile'
         Script OctopusDeployInstall
         {
             SetScript = {
-                Get-Service OctopusDeploy -ErrorAction Ignore | Stop-Service -Force
+                Get-Service OctopusDeploy -ErrorAction Ignore | Stop-Service OctopusDeploy -Force -Verbose | Write-Verbose
                 $isUpgrade = $false
                 if (Test-Path $using:octopusInstallStateFile) {
-                    $isUpgrade = $true
                     $currentVersion = [System.IO.FIle]::ReadAllText($using:octopusInstallStateFile).Trim()
-                    $currentOctopusInstallFile = Join-Path $using:octopusDeployRoot "OctopusServer.$currentVersion.msi"
-                     if (!(Test-Path $currentOctopusInstallFile)) {
-                         throw "Unable to install different Octopus Deploy versiom, previously installed msi file not found: $currentOctopusInstallFile"
-                     }
+                    if ($currentVersion -ne $using:OctopusVersionToInstall) {
+                        Write-Verbose 'Uninstalling current Octopus install'
+                        $isUpgrade = $true
+                        $currentOctopusInstallFile = Join-Path $using:octopusDeployRoot "OctopusServer.$currentVersion.msi"
+                        if (!(Test-Path $currentOctopusInstallFile)) {
+                            throw "Unable to install different Octopus Deploy versiom, previously installed msi file not found: $currentOctopusInstallFile"
+                        }
 
-                     $octopusUninstallLogFile = Join-Path $octopusDeployRoot "OctopusServer.$currentVersion.uninstall.log"
-                    $msiExitCode = (Start-Process -FilePath "msiexec.exe" -ArgumentList "/x `"$currentOctopusInstallFile`" /quiet /l*v `"$octopusUninstallLogFile`"" -Wait -Passthru).ExitCode
-                    if ($msiExitCode -ne 0)
-                    {
-                        throw "Uninstallation of Octopus Deploy failed; MSIEXEC exited with code: $msiExitCode"
+                        $octopusUninstallLogFile = Join-Path $octopusDeployRoot "OctopusServer.$currentVersion.uninstall.log"
+                        $msiExitCode = (Start-Process -FilePath "msiexec.exe" -ArgumentList "/x `"$currentOctopusInstallFile`" /quiet /l*v `"$octopusUninstallLogFile`"" -Wait -Passthru).ExitCode
+                        if ($msiExitCode -ne 0)
+                        {
+                            throw "Uninstallation of Octopus Deploy failed; MSIEXEC exited with code: $msiExitCode"
+                        }
+                        Remove-Item -Path $using:octopusInstallStateFile -Force
                     }
-                    Remove-Item -Path $using:octopusInstallStateFile -Force
                 }
+                Write-Verbose 'Starting Octopus MSI installer..'
                 $msiExitCode = (Start-Process -FilePath "msiexec.exe" -ArgumentList "/i `"$($using:octopusInstallFile)`"  /quiet /l*v `"$($using:octopusInstallLogFile)`"" -Wait -Passthru).ExitCode
                 if ($msiExitCode -ne 0)
                 {
                     throw "Installation of Octopus Deploy failed; MSIEXEC exited with code: $msiExitCode"
                 }
                 if ($isUpgrade) {
-                    Start-Service OctopusDeploy
+                   Start-Service OctopusDeploy -Verbose | Write-Verbose
                 }
                 [System.IO.FIle]::WriteAllText($using:octopusInstallStateFile, $using:OctopusVersionToInstall,[System.Text.Encoding]::ASCII)
             }
@@ -86,34 +95,94 @@ Configuration OctopusDeploy
             GetScript = { @{} }
             DependsOn = '[xRemoteFile]OctopusServer'
         }
-        $octopusConfigStateFile = Join-Path $octopusDeployRoot 'OctopusDeploy.config'
+
+        $octopusConfigStateFile = Join-Path $octopusDeployRoot 'configuration.statefile'
         $octopusConfigLogFile = Join-Path $octopusDeployRoot "OctopusServer.$OctopusVersionToInstall.config.log"
         Script OctopusDeployConfiguration
         {
             SetScript = {
                 $octopusServerExe = Join-Path $env:ProgramFiles 'Octopus Deploy\Octopus\Octopus.Server.exe'
-                & $octopusServerExe create-instance --console --instance OctopusServer --config "C:\Octopus\OctopusServer.config" *>> $using:octopusConfigLogFile
-                if ($LASTEXITCODE -gt 0) { throw "Exit code $LASTEXITCODE from Octopus Server: create-instance" }
-                & $octopusServerExe configure --console --instance OctopusServer --home "C:\Octopus" --storageConnectionString $using:ConnectionString --upgradeCheck "True" --upgradeCheckWithStatistics "True" --webAuthenticationMode "UsernamePassword" --webForceSSL "False" --webListenPrefixes $using:HostHeader --commsListenPort "10943" --serverNodeName $using:OctopusNodeName *>> $using:octopusConfigLogFile
-                if ($LASTEXITCODE -gt 0) { throw "Exit code $LASTEXITCODE from Octopus Server: configure" }
-                & $octopusServerExe database --console --instance OctopusServer --create *>> $using:octopusConfigLogFile
-                if ($LASTEXITCODE -gt 0) { throw "Exit code $LASTEXITCODE from Octopus Server: database" }
-                
-                $response = Invoke-WebRequest -UseBasicParsing -Uri "https://octopusdeploy.com/api/licenses/trial" -Method POST -Body @{ FullName=$env:USERNAME; Organization=$env:USERDOMAIN; EmailAddress="${env:USERNAME}@${env:USERDOMAIN}.com"; Source="azure" }
-                $licenseBase64 = [System.Convert]::ToBase64String(((New-Object System.Text.UTF8Encoding($false)).GetBytes($response.Content)))
-                & $octopusServerExe license --console --instance OctopusServer --licenseBase64 $licenseBase64 *>> $using:octopusConfigLogFile
-                if ($LASTEXITCODE -gt 0) { throw "Exit code $LASTEXITCODE from Octopus Server: license" }
 
-                & $octopusServerExe service --console --instance OctopusServer --install --reconfigure --start *>> $using:octopusConfigLogFile
-                if ($LASTEXITCODE -gt 0) { throw "Exit code $LASTEXITCODE from Octopus Server: service" }
-                Start-Service OctopusDeploy *>> $using:octopusConfigLogFile
+                & $octopusServerExe create-instance --console --instance OctopusServer --config "$($env:SystemDrive)\Octopus\OctopusServer.config" *>&1  | Write-Verbose
+                if ($LASTEXITCODE -ne 0) { throw "Exit code $LASTEXITCODE from Octopus Server: create-instance" }
+                & $octopusServerExe configure --console --instance OctopusServer --home "C:\Octopus" --storageConnectionString $using:ConnectionString --upgradeCheck "True" --upgradeCheckWithStatistics "True" --webAuthenticationMode "UsernamePassword" --webForceSSL "False" --webListenPrefixes $using:HostHeader --commsListenPort "10943" --serverNodeName $using:OctopusNodeName *>&1   | Write-Verbose
+                if ($LASTEXITCODE -ne 0) { throw "Exit code $LASTEXITCODE from Octopus Server: configure" }
+                & $octopusServerExe database --console --instance OctopusServer --create *>&1   | Write-Verbose
+                if ($LASTEXITCODE -ne 0) { throw "Exit code $LASTEXITCODE from Octopus Server: database" }
+                
+                $response = Invoke-WebRequest -UseBasicParsing -Uri "https://octopusdeploy.com/api/licenses/trial" -Method POST -Body @{ FullName=$env:USERNAME; Organization=$env:USERDOMAIN; EmailAddress="${env:USERNAME}@${env:USERDOMAIN}.onmicrosoft.com"; Source="azure" } -Verbose
+                $licenseBase64 = [System.Convert]::ToBase64String(((New-Object System.Text.UTF8Encoding($false)).GetBytes($response.Content)))
+                & $octopusServerExe license --console --instance OctopusServer --licenseBase64 $licenseBase64 *>&1   | Write-Verbose
+                if ($LASTEXITCODE -ne 0) { throw "Exit code $LASTEXITCODE from Octopus Server: license" }
+
+                & $octopusServerExe service --console --instance OctopusServer --install --reconfigure --stop *>&1   | Write-Verbose
+                if ($LASTEXITCODE -ne 0) { throw "Exit code $LASTEXITCODE from Octopus Server: service" }
+
                 [System.IO.FIle]::WriteAllText($using:octopusConfigStateFile, $LASTEXITCODE,[System.Text.Encoding]::ASCII)
             }
             TestScript = {
                 ((Test-Path $using:octopusConfigStateFile) -and ([System.IO.FIle]::ReadAllText($using:octopusConfigStateFile).Trim()) -eq '0')
             }
             GetScript = { @{} }
-            DependsOn = @('[xFirewall]OctopusDeployServer','[Script]OctopusDeployInstall')
+            DependsOn = '[Script]OctopusDeployInstall'
         }
+
+        $octopusServiceAccountUsername = $octopusDeployServiceAccount.UserName
+        User OctopusDeployServiceAccount
+        {
+            UserName                = $octopusServiceAccountUsername
+            Password                = $octopusDeployServiceAccount
+            PasswordChangeRequired  = $false
+            PasswordNeverExpires    = $true
+        }
+        Service OctopusDeploy
+        {
+            Name        = 'OctopusDeploy'
+            Credential  = $octopusDeployServiceAccount
+            StartupType = 'Automatic'
+            DependsOn = @('[User]OctopusDeployServiceAccount','[Script]OctopusDeployConfiguration')
+        } 
+
+        $octopusUrlAclStateFile = Join-Path $octopusDeployRoot 'urlacl.statefile'
+        Script URLAccessControlList
+        {
+            SetScript = {
+                $netsh = Join-Path -Resolve ([System.Environment]::SystemDirectory) 'netsh.exe'
+                Write-Verbose "Found $netsh"
+
+                Write-Verbose 'Removing existing URL access control entries for HTTP'
+                & $netsh http delete urlacl url=http://+:80/ *>&1 |  Write-Verbose
+                if ($LASTEXITCODE -notin @(0, 1)) { throw "Exit code $LASTEXITCODE from netsh delete urlacl" }
+                
+                $username = '{0}\{1}' -f [System.Environment]::MachineName, $using:octopusServiceAccountUsername
+                Write-Verbose "Adding new HTTP URL acccess control entry for user $username"
+                & $netsh http add urlacl url=$using:FullyQualifiedUrl user=$username *>&1 |  Write-Verbose
+                if ($LASTEXITCODE -ne 0) { throw "Exit code $LASTEXITCODE from netsh add urlacl" }
+                
+                [System.IO.FIle]::WriteAllText($using:octopusUrlAclStateFile, $LASTEXITCODE,[System.Text.Encoding]::ASCII)
+            }
+            TestScript = {
+                ((Test-Path $using:octopusUrlAclStateFile) -and ([System.IO.FIle]::ReadAllText($using:octopusUrlAclStateFile).Trim()) -eq '0')
+            }
+            GetScript = { @{} }
+            DependsOn = '[User]OctopusDeployServiceAccount'
+        }
+
+        $octopusServiceStartedStateFile = Join-Path $octopusDeployRoot 'service.statefile'
+        Script OctopusServiceStart
+        {
+            SetScript = {
+                Stop-Service OctopusDeploy -Force -Verbose | Write-Verbose
+                Start-Service OctopusDeploy -Verbose | Write-Verbose
+
+                [System.IO.FIle]::WriteAllText($using:octopusServiceStartedStateFile, (Get-Service OctopusDeploy | % Status),[System.Text.Encoding]::ASCII)
+            }
+            TestScript = {
+                ((Test-Path $using:octopusServiceStartedStateFile) -and ([System.IO.FIle]::ReadAllText($using:octopusServiceStartedStateFile).Trim()) -eq 'Running')
+            }
+            GetScript = { @{} }
+            DependsOn = @('[xFirewall]OctopusDeployServer','[Script]URLAccessControlList','[Service]OctopusDeploy','[Script]OctopusDeployConfiguration')
+        }
+
     }
 }
